requirements.txt
transformers
torch
Pillow
requests
accelerate
beautifulsoup4
